Added windows visual studio solution containing the exact code in the Ubuntu container. Refacter code to use doubles instead of floats

+++ Container/src/Utilities.cpp
@@ -1,21 +1,21 @@
 #include "Utilities.h"
 #include <cmath>
 
-std::vector<float> Utilities::createMatrixWave(int x, int y, int z, float frequency, float amplitude, float phase) {
-    std::vector<float> matrix(x * y * z);
+std::vector<double> Utilities::createMatrixWave(int x, int y, int z, double frequency, double amplitude, double phase) {
+    std::vector<double> matrix(x * y * z);
     for (int i = 0; i < x * y * z; i++) {
-        float waveValue = amplitude * sin(frequency * i + phase);
+        double waveValue = amplitude * sin(frequency * i + phase);
         matrix[i] = waveValue;
     }
     return matrix;
 }
 
-std::vector<float> Utilities::createMatrixRandom(int x, int y, int z, float minVal, float maxVal) {
+std::vector<double> Utilities::createMatrixRandom(int x, int y, int z, double minVal, double maxVal) {
     std::random_device rd;
     std::mt19937 gen(rd());
     std::uniform_real_distribution<> dis(minVal, maxVal);
 
-    std::vector<float> matrix(x * y * z);
+    std::vector<double> matrix(x * y * z);
     for (int i = 0; i < x * y * z; i++) {
         matrix[i] = dis(gen);
     }
