+++ Container/src/Utilities.h
@@ -3,7 +3,7 @@
 
 class Utilities {
 public:
-    static std::vector<float> createMatrixWave(int x, int y, int z, float frequency, float amplitude, float phase);
-    static std::vector<float> createMatrixRandom(int x, int y, int z, float minVal, float maxVal);
+    static std::vector<double> createMatrixWave(int x, int y, int z, double frequency, double amplitude, double phase);
+    static std::vector<double> createMatrixRandom(int x, int y, int z, double minVal, double maxVal);
 };
